+++ src/main/java/domain/coordination/Column.java
@@ -3,9 +3,10 @@ package domain.coordination;
 import java.util.ArrayList;
 import java.util.List;
 
-import static util.ErrorMessage.INVALID_COORDINATION;
 
 public record Column(int index) {
+
+    private static final String INVALID_COORDINATION = "좌표값이 잘못되었습니다.";
 
     private static final int MIN = 1;
     private static final int MAX = 9;
@@ -16,7 +17,7 @@ public record Column(int index) {
 
     private void validateRange(int index) {
         if (index < MIN || index > MAX) {
-            throw new IllegalArgumentException(INVALID_COORDINATION.getMessage());
+            throw new IllegalArgumentException(INVALID_COORDINATION);
         }
     }
