+++ src/main/java/domain/coordination/Row.java
@@ -1,11 +1,11 @@
 package domain.coordination;
 
-import util.ErrorMessage;
-
 import java.util.ArrayList;
 import java.util.List;
 
 public record Row(int index) {
+
+    private static final String INVALID_COORDINATION = "좌표값이 잘못되었습니다.";
 
     private static final int MIN = 1;
     private static final int MAX = 10;
@@ -24,7 +24,7 @@ public record Row(int index) {
 
     private void validateRange(int index) {
         if (index < MIN || index > MAX) {
-            throw new IllegalArgumentException(ErrorMessage.INVALID_COORDINATION.getMessage());
+            throw new IllegalArgumentException(INVALID_COORDINATION);
         }
     }
