+++ src/main/java/domain/piece/EmptyPiece.java
@@ -2,12 +2,11 @@ package domain.piece;
 
 import domain.coordination.Coordination;
 import domain.piece.error.PieceException;
-
 import java.util.Map;
 
-import static util.ErrorMessage.NOT_EXISTS_PIECE;
-
 public class EmptyPiece extends Piece {
+
+    private static final String NOT_EXISTS_PIECE = "기물이 존재하지 않습니다.";
 
     public EmptyPiece(Team team) {
         super(team);
@@ -25,6 +24,6 @@ public class EmptyPiece extends Piece {
 
     @Override
     public void validateMovable(Coordination from, Coordination to, Map<Coordination, Piece> board) {
-        throw new PieceException(NOT_EXISTS_PIECE.getMessage());
+        throw new PieceException(NOT_EXISTS_PIECE);
     }
 }
