+++ src/main/java/domain/piece/Team.java
@@ -2,14 +2,14 @@ package domain.piece;
 
 import domain.game.Turn;
 
-import static util.ErrorMessage.NOT_SAME_TEAM;
-
 public enum Team {
 
     CHO(true),
     HAN(false),
     NONE(false),
     ;
+
+    private static final String NOT_SAME_TEAM = "본인의 진영의 기물이 아닙니다.";
 
     private final boolean isCho;
 
@@ -23,7 +23,7 @@ public enum Team {
 
     public void validateSameTeam(Turn turn) {
         if (this.isCho != turn.isCho()) {
-            throw new IllegalArgumentException(NOT_SAME_TEAM.getMessage());
+            throw new IllegalArgumentException(NOT_SAME_TEAM);
         }
     }
 }
