+++ src/main/java/util/InputParser.java
@@ -7,9 +7,12 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static util.ErrorMessage.*;
-
 public class InputParser {
+
+    private static final String ERROR_BLANK_INPUT = "입력값이 비어 있습니다. '열,행' 형식으로 입력하세요. (예: 5,1)";
+    private static final String ERROR_TOKEN_COUNT = "입력값은 쉼표(,)로 구분된 2개여야 합니다. (입력된 값 %d개) (예: 5,1)";
+    private static final String ERROR_BLANK_TOKEN = "%d번째 값이 비어 있습니다. '열,행' 형식으로 입력하세요. (예: 5,1)";
+    private static final String ERROR_NUMBER_FORMAT = "'%s'는 숫자가 아닙니다. 숫자만 입력하세요. (예: 5,1)";
 
     private static final Pattern NUMBER_PATTERN = Pattern.compile("^\\d+$");
     private static final String DELIMITER = ",";
@@ -29,7 +32,7 @@ public class InputParser {
 
     private static void validateNotBlank(String input) {
         if (input == null || input.isBlank()) {
-            throw new IllegalArgumentException(ERROR_BLANK_INPUT.getMessage());
+            throw new IllegalArgumentException(ERROR_BLANK_INPUT);
         }
     }
 
@@ -41,7 +44,7 @@ public class InputParser {
 
     private static void validateTokenCount(List<String> tokens) {
         if (tokens.size() != REQUIRED_TOKEN_COUNT) {
-            throw new IllegalArgumentException(String.format(ERROR_TOKEN_COUNT.getMessage(), tokens.size()));
+            throw new IllegalArgumentException(String.format(ERROR_TOKEN_COUNT, tokens.size()));
         }
     }
 
@@ -51,7 +54,7 @@ public class InputParser {
                 .findFirst()
                 .ifPresent(i -> {
                     throw new IllegalArgumentException(
-                            String.format(ERROR_BLANK_TOKEN.getMessage(), i + 1)
+                            String.format(ERROR_BLANK_TOKEN, i + 1)
                     );
                 });
     }
@@ -59,7 +62,7 @@ public class InputParser {
     private static void validateNumberFormat(List<String> tokens) {
         for (String token : tokens) {
             if (!NUMBER_PATTERN.matcher(token).matches()) {
-                throw new IllegalArgumentException(String.format(ERROR_NUMBER_FORMAT.getMessage(), token));
+                throw new IllegalArgumentException(String.format(ERROR_NUMBER_FORMAT, token));
             }
         }
     }
